Move folder d from app/d/(dashboard) to app/(dashboard)/d
